refactor: Add enums package
