+++ api/src/main/java/grooteogi/domain/User.java
@@ -1,5 +1,6 @@
 package grooteogi.domain;
 
+import grooteogi.enums.LoginType;
 import javax.persistence.*;
 
 import lombok.Getter;
