+++ api/src/main/java/grooteogi/enums/HashtagType.java
@@ -1,4 +1,4 @@
-package grooteogi.domain;
+package grooteogi.enums;
 
 public enum HashtagType {
     CONCERN,
