+++ api/src/main/java/grooteogi/enums/LoginType.java
@@ -1,4 +1,4 @@
-package grooteogi.domain;
+package grooteogi.enums;
 
 public enum LoginType {
     GENERAL, KAKAO, FACEBOOK
